Refactoriza estructura del timeline
